requirements.txt
alembic==1.12.1
bleach==6.1.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
dominate==2.8.0
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
